requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
Brotli @ file:///croot/brotli-split_1714483155106/work
cattrs==24.1.2
certifi @ file:///croot/certifi_1720453481653/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock @ file:///croot/filelock_1700591183607/work
Flask==3.0.3
flexcache==0.3
flexparser==0.3.1
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.27.0
idna @ file:///croot/idna_1714398848350/work
igraph==0.11.8
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
littleutils==0.2.4
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
molesq==0.4.0
morphops==0.1.13
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.5
narwhals==1.26.0
navis==1.9.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
ncollpyde==0.19.0
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1720002482208/work
neuprint-python==0.4.26
notebook==7.2.2
notebook_shim==0.2.4
nptyping==2.5.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
ogb==1.3.6
onnx==1.17.0
outdated==0.2.2
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow @ file:///croot/pillow_1721059439630/work
Pint==0.24.3
platformdirs==4.2.2
plotly==6.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.28.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pyg-lib==0.4.0+pt23cu121
Pygments==2.18.0
pykdtree==1.3.13
pynrrd==1.0.0
pyparsing==3.1.4
pypng==0.20220715.0
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq==26.2.0
rdata==0.11.2
referencing==0.35.1
requests @ file:///croot/requests_1721410876868/work
requests-cache==1.2.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
safetensors==0.5.0
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.18.0
setproctitle==1.3.3
six==1.16.0
skeletor==1.3.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy @ file:///croot/sympy_1724938189289/work
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.5.0
timm==0.6.5
tinycss2==1.3.0
tomli==2.0.1
torch==2.3.0+cu121
torch_cluster==1.6.3+pt23cu121
torch_geometric==2.5.3
torch_scatter==2.1.2+pt23cu121
torch_sparse==0.6.18+pt23cu121
torch_spline_conv==1.2.2+pt23cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trimesh==4.5.1
triton==2.3.0
types-python-dateutil==2.9.0.20240906
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
url-normalize==1.4.3
urllib3 @ file:///croot/urllib3_1718912636303/work
wandb==0.18.6
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
xarray==2024.10.0
yarl==1.9.8
zipp==3.21.0
